refactor home page (Up to get involved)
